requirements.txt
streamlit
google-generativeai
qdrant-client
PyPDF2
python-dotenv
langchain
langchain-text-splitters
numpy
pandas
sentence-transformers
torch
transformers